requirements.txt
django==5.2.8
django-environ==0.12.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
django-storages==1.14.6
django-ckeditor==6.7.3
psycopg2==2.9.11